requirement.txt
streamlit
textblob==0.17.1
nltk==3.8.1
google-generativeai
python-dotenv